﻿Add Action to Build and Deploy to Itch (#136)
This commit adds a github action that builds a
unity project across multiple targets (Android and WebGL), and
uploads the build artifacts, and pushes them to Itch.io. The action
comes with multiple parameters such as the environment, build name,
and the version tag.

I have tried my best to make this very extensible and portable too. It
should be easy to add new targets.

TODOs:
- Bundle the license file in the zip artifacts